Flatpak distribution of Climaxima
Climaxima is a user interface built on top of Maxima, a CAS based
on Macsyma which was started in the 1960's.

The application has the ability to perform regular algebra functions,
do integration and differentiation, solve differential equations and
a lot of other things.

The Climaxima interface adds proper mathematical rendering of
equations as well as other features to  make using Maxima much
easier.

Climaxima is still a work in process, but it has reached a point where
it can be useful for more people.

The project is hosted here: https://github.com/lokedhs/maxima-client

For more information about Maxima: http://maxima.sourceforge.net/